+++ web/assets/js/parsley-validator.js
@@ -31,3 +31,23 @@ window.Parsley.addValidator('lt', {
   },
   priority: 32
 });
+
+var validateDateFormat = function (value) {
+  var parts = value.split('/');
+  if (parts.length < 3) {
+    return false;
+  }
+  var dt = new Date(parts[2], parts[0] - 1, parts[1]);
+  return (dt && dt.getMonth() === parseInt(parts[0], 10) - 1 && dt.getFullYear() === parseInt(parts[2]));
+};
+
+// Date format validator
+window.Parsley.addValidator('dateMdy', {
+  validateString: function (value) {
+    return validateDateFormat(value);
+  },
+  messages: {
+    en: 'Invalid date format.'
+  },
+  priority: 32
+});
